Normalize attachment access identifiers

+++ admin/code/tce_attachment.php
@@ -1,7 +1,7 @@
 <?php
 
 require_once '../config/tce_config.php';
-$pagelevel = K_AUTH_ADMIN_RESULTS;
+$pagelevel = (int) K_AUTH_ADMIN_RESULTS;
 require_once '../../shared/code/tce_authorization.php';
 require_once '../../shared/code/tce_functions_attachments.php';
